+++ packages/vite/src/node/ssr/ssrExternal.ts
@@ -5,6 +5,7 @@ import {
   createDebugger,
   isDefined,
   lookupFile,
+  normalizePath,
   resolveFrom,
   unique
 } from '../utils'
@@ -68,7 +69,9 @@ export function resolveSSRExternal(
         undefined,
         true
       )?.id
-      requireEntry = require.resolve(id, { paths: [root] })
+      // normalizePath required for windows. tryNodeResolve uses normalizePath
+      // which returns with '/', require.resolve returns with '\\'
+      requireEntry = normalizePath(require.resolve(id, { paths: [root] }))
     } catch (e) {
       // resolve failed, assume include
       debug(`Failed to resolve entries for package "${id}"\n`, e)